2023 Day 10
Currently hardcoded because in my input 'S' is a '|' character. This
solution may not work with all inputs.

I'm happy with how I solved part two, but not happy with how long it
took to debug. My algorithm works just fine, but I completely forgot to
consider the 'S' in the grid.